Give hint about public prefix for env variables in browser
I lost quite a few hours figuring out why an environment variable was `undefined` in the browser. I only read the part where it says that there is built-in support for `.env` files if you append them with `.local`. I never reached the part where it says you have to prefix it with `NEXT_PUBLIC_` in order to expose those variables to the browser.
I think with these changes, by giving the hint further up the page, it will save future people who are like me some time and a headache.